pyproject.toml
[tool.poetry]
name = "federated-learning-dev"
version = "0.1.0"
description = ""
authors = ["Maximilian Forstenhaeusler  <maxi.forstenhaeusler@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
# Use newer python versions at your own risk, they have not been fully tested with `pfl`.
python = ">=3.10,<3.11"
awscli = "^1.32.29"
h5py = "^3.8.0"
multiprocess = "^0.70.12"
tqdm = "^4.63.1"
tensorflow = { version = "^2.14.0", optional = true }
tensorflow_probability = { version = "^0.22", optional = true }
tensorflow_addons = { version = ">=0.20.0,<1", optional = true }
torch = [
  { version = "^2.4.1+cu118", source = "torch_cu118", markers = "sys_platform == 'linux'", optional = true },
  { version = "^2.4.1", source = "PyPI", markers = "sys_platform == 'darwin'",  optional = true },
]
torchvision = [
  { version = "^0.19.1+cu118", source = "torch_cu118", markers = "sys_platform == 'linux'", optional = true },
  { version = "^0.19.1", source = "PyPI", markers = "sys_platform == 'darwin'",  optional = true },
]
# Installs pfl from source.
pfl = {extras = ["pytorch"], version = "^0.2.0"}
pillow = ">=10.2.0"
datasets = "^2.19.1"
# Fixes vulnerability https://github.com/advisories/GHSA-2g68-c3qc-8985
werkzeug = ">=3.0.3"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"
wandb = "^0.18.5"
matplotlib = "^3.9.2"
pandas = "^2.2.3"
jupyterlab = "^4.2.5"
ujson = "^5.10.0"
boto3 = "^1.35.50"
awswrangler = "^3.9.1"
hydra-core = "^1.3.2"
plotly = "^5.24.1"
expecttest = "^0.3.0"
hypothesis = "^6.123.9"
horovod = {extras = ["pytorch"], version = "^0.28.1"}


[tool.poetry.extras]
tf = ["pfl", "tensorflow_addons", "tensorflow"]
pytorch = ["pfl", "torchvision", "torch"]
mlx = ["pfl", "mlx"]

[tool.poetry.group.dev.dependencies]
mock = "^5.1.0"
pytest = "^7.2.0"
pre-commit = "^2.20.0"
pytest-xdist = "^3.3.1"
pytest-lazy-fixture = "^0.6.3"
mypy = "1.5"
ruff = "0.0.290"
scikit-learn = "^1.0.2"
yapf = "^0.40.1"

# TODO: these pytorch sources are only needed while pfl is installed from github address.
[[tool.poetry.source]]
name = "torch_cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "supplemental"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.mypy]
disable_error_code = "import,union-attr,index,attr-defined,arg-type"


